SpriteBatch access memory optimisation
Optimized CPU usage while processing with the SpriteBatch. This commit
decreases the cpu and gpu usage of the app by about 30%. Basically it
doesn't recreate a SpriteBatch at each frame, but each time the user
resises the app. This should be extremely powerful on android.

+++ core/src/com/norwand/game/management/gamedata/environement/FloorData.java
@@ -69,7 +69,7 @@ public class FloorData {
 	tiles = new Tile[width * height];
 	for (int i = 0; i < width * height; i++) {
 	    try {
-		tiles[i] = new Tile(converttiletype(phy[i] - 1),
+		tiles[i] = new Tile((byte) converttiletype(phy[i] - 1),
 			new AnimatedSprite[] {
 				ImagesHolder.tileset
 					.getTileAnimation(bg1[i] - 1),
