+++ core/src/com/norwand/game/management/gamedata/environement/tiles/Tile.java
@@ -4,7 +4,7 @@ import com.norwand.game.utility.objects.AnimatedSprite;
 
 public class Tile {
 
-    public Tile(int type, AnimatedSprite[] background,
+    public Tile(byte type, AnimatedSprite[] background,
 	    AnimatedSprite[] foreground) {
 	this.background = background;
 	this.foreground = foreground;
@@ -12,14 +12,14 @@ public class Tile {
     }
 
     /** The physics of the tile. */
-    public int type;
-    public static final int TYPE_NORMAL = 0;
-    public static final int TYPE_SOLID = 1;
-    public static final int TYPE_ICE = 2;
-    public static final int TYPE_WATER = 3;
-    public static final int TYPE_LAVA = 4;
-    public static final int TYPE_ACID = 5;
-    public static final int TYPE_VOID = 6;
+    public byte type;
+    public static final byte TYPE_NORMAL = 0;
+    public static final byte TYPE_SOLID = 1;
+    public static final byte TYPE_ICE = 2;
+    public static final byte TYPE_WATER = 3;
+    public static final byte TYPE_LAVA = 4;
+    public static final byte TYPE_ACID = 5;
+    public static final byte TYPE_VOID = 6;
 
     public AnimatedSprite[] background;
     public AnimatedSprite[] foreground;
